requirements.txt
python-dotenv
langchain
anthropic
langchain_anthropic
langchain_openai
crewai
pyautogen
pytest
pytest-mock
pytest-cov
embedchain